% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_history.R
\name{stop_motion}
\alias{stop_motion}
\title{Show previous plots as a stop motion slideshow}
\usage{
stop_motion(
  height = 500,
  dir = GG_RECORDING_ENV$recording_dir,
  ext = GG_RECORDING_ENV$device_ext,
  filename = NULL
)
}
\arguments{
\item{height}{Set the height of the carousel widget and images in pixels. Defaults to 500.}

\item{dir}{directory the saved intermediate plots are in. When able, uses the set directory from `gg_record`.}

\item{ext}{extension type of the saved intermediate plots to display. When able, uses the set device extention from `gg_record`.}
}
\value{
Returns a slickR htmlwidget populated with the plots
}
\description{
Using slickjs, preview all historical plots as a image carousel using slick js via the
slickR package.
}
\details{
slickR and htmltools must be installed for this function to operate.
}
\examples{

if(require(ggplot2) & interactive()){

  gg_record(dir = file.path(tempdir(),"recording"), device = "png" )
  ggplot(data.frame(x = 1, y = 1), aes(x=x, y=y)) + geom_point() + ylim(0,4)
  ggplot(data.frame(x = 1, y = 2), aes(x=x, y=y)) + geom_point() + ylim(0,4)

  ## resize canvas of the last plot
  gg_resize_film(height = 10, width = 5, dpi = 350)

  ggplot(data.frame(x = 1, y = 3), aes(x=x, y=y)) + geom_point() + ylim(0,4)

  stop_motion()
}


}
